Rewriting plugin for Terminus v1.x
